Integrate start (#2262)
* integrate start

* integrate start

* update etna network

* integrate start

* integrate start

* finish integrate start

* fix lint

* stop existing local avalanchego

* more robust
