+++ cmd/nodecmd/local.go
@@ -3,21 +3,10 @@
 package nodecmd
 
 import (
-	"fmt"
-	"github.com/ava-labs/avalanche-cli/pkg/node"
-	"os"
-	"path/filepath"
-
-	"github.com/ava-labs/avalanche-cli/pkg/binutils"
 	"github.com/ava-labs/avalanche-cli/pkg/cobrautils"
-	"github.com/ava-labs/avalanche-cli/pkg/constants"
-	"github.com/ava-labs/avalanche-cli/pkg/localnet"
-	"github.com/ava-labs/avalanche-cli/pkg/models"
 	"github.com/ava-labs/avalanche-cli/pkg/networkoptions"
-	"github.com/ava-labs/avalanche-cli/pkg/subnet"
-	"github.com/ava-labs/avalanche-cli/pkg/utils"
+	"github.com/ava-labs/avalanche-cli/pkg/node"
 	"github.com/ava-labs/avalanche-cli/pkg/ux"
-	"github.com/ava-labs/avalanche-network-runner/client"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/spf13/cobra"
 )
@@ -121,341 +110,33 @@ func newLocalDestroyCmd() *cobra.Command {
 	}
 }
 
-// stub for now
-func preLocalChecks(clusterName string) error {
-	clusterExists, err := checkClusterExists(clusterName)
-	if err != nil {
-		return fmt.Errorf("error checking cluster: %w", err)
-	}
-	if clusterExists {
-		return fmt.Errorf("cluster %q already exists", clusterName)
-	}
-	// expand passed paths
-	if genesisPath != "" {
-		genesisPath = utils.ExpandHome(genesisPath)
-	}
-	if upgradePath != "" {
-		upgradePath = utils.ExpandHome(upgradePath)
-	}
-	// checks
-	if useCustomAvalanchegoVersion != "" && (useLatestAvalanchegoReleaseVersion || useLatestAvalanchegoPreReleaseVersion) {
-		return fmt.Errorf("specify either --custom-avalanchego-version or --latest-avalanchego-version")
-	}
-	if avalanchegoBinaryPath != "" && (useLatestAvalanchegoReleaseVersion || useLatestAvalanchegoPreReleaseVersion || useCustomAvalanchegoVersion != "") {
-		return fmt.Errorf("specify either --avalanchego-path or --latest-avalanchego-version or --custom-avalanchego-version")
-	}
-	if useEtnaDevnet && (globalNetworkFlags.UseDevnet || globalNetworkFlags.UseFuji) {
-		return fmt.Errorf("etna devnet can only be used with devnet")
-	}
-	if useEtnaDevnet && genesisPath != "" {
-		return fmt.Errorf("etna devnet uses predefined genesis file")
-	}
-	if useEtnaDevnet && upgradePath != "" {
-		return fmt.Errorf("etna devnet uses predefined upgrade file")
-	}
-	if useEtnaDevnet && (len(bootstrapIDs) != 0 || len(bootstrapIPs) != 0) {
-		return fmt.Errorf("etna devnet uses predefined bootstrap configuration")
-	}
-	if len(bootstrapIDs) != len(bootstrapIPs) {
-		return fmt.Errorf("number of bootstrap IDs and bootstrap IP:port pairs must be equal")
-	}
-	if genesisPath != "" && !utils.FileExists(genesisPath) {
-		return fmt.Errorf("genesis file %s does not exist", genesisPath)
-	}
-	if upgradePath != "" && !utils.FileExists(upgradePath) {
-		return fmt.Errorf("upgrade file %s does not exist", upgradePath)
-	}
-	return nil
-}
-
-func localClusterDataExists(clusterName string) bool {
-	rootDir := app.GetLocalDir(clusterName)
-	return utils.FileExists(filepath.Join(rootDir, "state.json"))
-}
-
 func localStartNode(_ *cobra.Command, args []string) error {
-	var err error
 	clusterName := args[0]
-	network := models.UndefinedNetwork
-
-	// check if this is existing cluster
-	rootDir := app.GetLocalDir(clusterName)
-	pluginDir := filepath.Join(rootDir, "node1", "plugins")
-	// make sure rootDir exists
-	if err := os.MkdirAll(rootDir, 0o700); err != nil {
-		return fmt.Errorf("could not create root directory %s: %w", rootDir, err)
+	anrSettings := node.ANRSettings{
+		GenesisPath:          genesisPath,
+		UpgradePath:          upgradePath,
+		BootstrapIDs:         bootstrapIDs,
+		BootstrapIPs:         bootstrapIPs,
+		StakingSignerKeyPath: stakingTLSKeyPath,
+		StakingCertKeyPath:   stakingCertKeyPath,
+		StakingTLSKeyPath:    stakingTLSKeyPath,
 	}
-	// make sure pluginDir exists
-	if err := os.MkdirAll(pluginDir, 0o700); err != nil {
-		return fmt.Errorf("could not create plugin directory %s: %w", pluginDir, err)
+	avaGoVersionSetting := node.AvalancheGoVersionSettings{
+		UseCustomAvalanchegoVersion:           useCustomAvalanchegoVersion,
+		UseLatestAvalanchegoPreReleaseVersion: useLatestAvalanchegoPreReleaseVersion,
+		UseLatestAvalanchegoReleaseVersion:    useLatestAvalanchegoReleaseVersion,
+		UseAvalanchegoVersionFromSubnet:       useAvalanchegoVersionFromSubnet,
 	}
-	ctx, cancel := utils.GetANRContext()
-	defer cancel()
-
-	// starts server
-	avalancheGoVersion := "latest"
-	if avalanchegoBinaryPath == "" {
-		avalancheGoVersion, err = getAvalancheGoVersion()
-		if err != nil {
-			return err
-		} else {
-			ux.Logger.PrintToUser("Using AvalancheGo version: %s", avalancheGoVersion)
-		}
-	}
-	serverLogPath := filepath.Join(rootDir, "server.log")
-	sd := subnet.NewLocalDeployer(app, avalancheGoVersion, avalanchegoBinaryPath, "")
-	if err := sd.StartServer(
-		constants.ServerRunFileLocalClusterPrefix,
-		binutils.LocalClusterGRPCServerPort,
-		binutils.LocalClusterGRPCGatewayPort,
-		rootDir,
-		serverLogPath,
-	); err != nil {
-		return err
-	}
-	_, avalancheGoBinPath, err := sd.SetupLocalEnv()
-	if err != nil {
-		return err
-	}
-	cli, err := binutils.NewGRPCClientWithEndpoint(binutils.LocalClusterGRPCServerEndpoint)
-	if err != nil {
-		return err
-	}
-	alreadyBootstrapped, err := localnet.CheckNetworkIsAlreadyBootstrapped(ctx, cli)
-	if err != nil {
-		return err
-	}
-	if alreadyBootstrapped {
-		ux.Logger.PrintToUser("")
-		ux.Logger.PrintToUser("A local cluster is already executing")
-		ux.Logger.PrintToUser("please stop it by calling 'node local stop'")
-		return nil
-	}
-
-	if localClusterDataExists(clusterName) {
-		ux.Logger.GreenCheckmarkToUser("Local cluster %s found. Booting up...", clusterName)
-		loadSnapshotOpts := []client.OpOption{
-			client.WithReassignPortsIfUsed(true),
-			client.WithPluginDir(pluginDir),
-			client.WithSnapshotPath(rootDir),
-		}
-		// load snapshot for existing network
-		if _, err := cli.LoadSnapshot(
-			ctx,
-			clusterName,
-			true, // in-place
-			loadSnapshotOpts...,
-		); err != nil {
-			return fmt.Errorf("failed to load snapshot: %w", err)
-		}
-	} else {
-		ux.Logger.GreenCheckmarkToUser("Local cluster %s not found. Creating...", clusterName)
-		if useEtnaDevnet {
-			network = models.NewNetwork(
-				models.Devnet,
-				constants.EtnaDevnetNetworkID,
-				constants.EtnaDevnetEndpoint,
-				clusterName,
-			)
-		} else {
-			network, err = networkoptions.GetNetworkFromCmdLineFlags(
-				app,
-				"",
-				globalNetworkFlags,
-				false,
-				true,
-				createSupportedNetworkOptions,
-				"",
-			)
-			if err != nil {
-				return err
-			}
-		}
-		if err := preLocalChecks(clusterName); err != nil {
-			return err
-		}
-		if useEtnaDevnet {
-			bootstrapIDs = constants.EtnaDevnetBootstrapNodeIDs
-			bootstrapIPs = constants.EtnaDevnetBootstrapIPs
-			// prepare genesis and upgrade files for anr
-			genesisFile, err := os.CreateTemp("", "etna_devnet_genesis")
-			if err != nil {
-				return fmt.Errorf("could not create save Etna Devnet genesis file: %w", err)
-			}
-			if _, err := genesisFile.Write(constants.EtnaDevnetGenesisData); err != nil {
-				return fmt.Errorf("could not write Etna Devnet genesis data: %w", err)
-			}
-			if err := genesisFile.Close(); err != nil {
-				return fmt.Errorf("could not close Etna Devnet genesis file: %w", err)
-			}
-			genesisPath = genesisFile.Name()
-			defer os.Remove(genesisPath)
-
-			upgradeFile, err := os.CreateTemp("", "etna_devnet_upgrade")
-			if err != nil {
-				return fmt.Errorf("could not create save Etna Devnet upgrade file: %w", err)
-			}
-			if _, err := upgradeFile.Write(constants.EtnaDevnetUpgradeData); err != nil {
-				return fmt.Errorf("could not write Etna Devnet upgrade data: %w", err)
-			}
-			upgradePath = upgradeFile.Name()
-			if err := upgradeFile.Close(); err != nil {
-				return fmt.Errorf("could not close Etna Devnet upgrade file: %w", err)
-			}
-			defer os.Remove(upgradePath)
-		}
-
-		if stakingTLSKeyPath != "" && stakingCertKeyPath != "" && stakingSignerKeyPath != "" {
-			if err := os.MkdirAll(filepath.Join(rootDir, "node1", "staking"), 0o700); err != nil {
-				return fmt.Errorf("could not create root directory %s: %w", rootDir, err)
-			}
-			if err := utils.FileCopy(stakingTLSKeyPath, filepath.Join(rootDir, "node1", "staking", "staker.key")); err != nil {
-				return err
-			}
-			if err := utils.FileCopy(stakingCertKeyPath, filepath.Join(rootDir, "node1", "staking", "staker.crt")); err != nil {
-				return err
-			}
-			if err := utils.FileCopy(stakingSignerKeyPath, filepath.Join(rootDir, "node1", "staking", "signer.key")); err != nil {
-				return err
-			}
-		}
-
-		anrOpts := []client.OpOption{
-			client.WithNumNodes(1),
-			client.WithNetworkID(network.ID),
-			client.WithExecPath(avalancheGoBinPath),
-			client.WithRootDataDir(rootDir),
-			client.WithReassignPortsIfUsed(true),
-			client.WithPluginDir(pluginDir),
-			client.WithFreshStakingIds(true),
-			client.WithZeroIP(false),
-		}
-		if genesisPath != "" && utils.FileExists(genesisPath) {
-			anrOpts = append(anrOpts, client.WithGenesisPath(genesisPath))
-		}
-		if upgradePath != "" && utils.FileExists(upgradePath) {
-			anrOpts = append(anrOpts, client.WithUpgradePath(upgradePath))
-		}
-		if bootstrapIDs != nil {
-			anrOpts = append(anrOpts, client.WithBootstrapNodeIDs(bootstrapIDs))
-		}
-		if bootstrapIPs != nil {
-			anrOpts = append(anrOpts, client.WithBootstrapNodeIPPortPairs(bootstrapIPs))
-		}
-
-		ux.Logger.PrintToUser("Starting local avalanchego node using root: %s ...", rootDir)
-		spinSession := ux.NewUserSpinner()
-		spinner := spinSession.SpinToUser("Booting Network. Wait until healthy...")
-		if _, err := cli.Start(ctx, avalancheGoBinPath, anrOpts...); err != nil {
-			ux.SpinFailWithError(spinner, "", err)
-			localDestroyNode(nil, []string{clusterName})
-			return fmt.Errorf("failed to start local avalanchego: %w", err)
-		}
-		ux.SpinComplete(spinner)
-		// save cluster config for the new local cluster
-		if err := addLocalClusterConfig(network); err != nil {
-			return err
-		}
-	}
-
-	ux.Logger.GreenCheckmarkToUser("Avalanchego started and ready to use from %s", rootDir)
-	ux.Logger.PrintToUser("")
-	ux.Logger.PrintToUser("Node logs directory: %s/node1/logs", rootDir)
-	ux.Logger.PrintToUser("")
-	ux.Logger.PrintToUser("Network ready to use.")
-	ux.Logger.PrintToUser("")
-
-	status, err := cli.Status(ctx)
-	if err != nil {
-		return err
-	}
-	for _, nodeInfo := range status.ClusterInfo.NodeInfos {
-		ux.Logger.PrintToUser("URI: %s", nodeInfo.Uri)
-		ux.Logger.PrintToUser("NodeID: %s", nodeInfo.Id)
-		ux.Logger.PrintToUser("")
-	}
-
-	return nil
+	return node.StartLocalNode(app, clusterName, useEtnaDevnet, avalanchegoBinaryPath, anrSettings, avaGoVersionSetting, globalNetworkFlags, createSupportedNetworkOptions)
 }
 
 func localStopNode(_ *cobra.Command, _ []string) error {
-	cli, err := binutils.NewGRPCClientWithEndpoint(
-		binutils.LocalClusterGRPCServerEndpoint,
-		binutils.WithAvoidRPCVersionCheck(true),
-		binutils.WithDialTimeout(constants.FastGRPCDialTimeout),
-	)
-	if err != nil {
-		return err
-	}
-	ctx, cancel := utils.GetANRContext()
-	defer cancel()
-	bootstrapped, err := localnet.CheckNetworkIsAlreadyBootstrapped(ctx, cli)
-	if err != nil {
-		return err
-	}
-	if bootstrapped {
-		if _, err = cli.Stop(ctx); err != nil {
-			return fmt.Errorf("failed to stop avalanchego: %w", err)
-		}
-	}
-	if err := binutils.KillgRPCServerProcess(
-		app,
-		binutils.LocalClusterGRPCServerEndpoint,
-		constants.ServerRunFileLocalClusterPrefix,
-	); err != nil {
-		return err
-	}
-	ux.Logger.GreenCheckmarkToUser("avalanchego stopped")
-	return nil
+	return node.StopLocalNode(app)
 }
 
 func localDestroyNode(_ *cobra.Command, args []string) error {
 	clusterName := args[0]
-
-	localStopNode(nil, nil)
-
-	rootDir := app.GetLocalDir(clusterName)
-	if err := os.RemoveAll(rootDir); err != nil {
-		return err
-	}
-
-	if ok, err := checkClusterIsLocal(clusterName); err != nil || !ok {
-		return fmt.Errorf("local cluster %q not found", clusterName)
-	}
-
-	clustersConfig, err := app.LoadClustersConfig()
-	if err != nil {
-		return err
-	}
-	delete(clustersConfig.Clusters, clusterName)
-	if err := app.WriteClustersConfigFile(&clustersConfig); err != nil {
-		return err
-	}
-
-	ux.Logger.GreenCheckmarkToUser("Local node %s cleaned up.", clusterName)
-	return nil
-}
-
-func addLocalClusterConfig(network models.Network) error {
-	clusterName := network.ClusterName
-	clustersConfig, err := app.LoadClustersConfig()
-	if err != nil {
-		return err
-	}
-	clusterConfig := clustersConfig.Clusters[clusterName]
-	clusterConfig.Local = true
-	clusterConfig.Network = network
-	clustersConfig.Clusters[clusterName] = clusterConfig
-	return app.WriteClustersConfigFile(&clustersConfig)
-}
-
-func checkClusterIsLocal(clusterName string) (bool, error) {
-	clustersConfig, err := app.GetClustersConfig()
-	if err != nil {
-		return false, err
-	}
-	clusterConf, ok := clustersConfig.Clusters[clusterName]
-	return ok && clusterConf.Local, nil
+	return node.DestroyLocalNode(app, clusterName)
 }
 
 func localTrack(_ *cobra.Command, args []string) error {
@@ -465,12 +146,4 @@ func localTrack(_ *cobra.Command, args []string) error {
 func notImplementedForLocal(what string) error {
 	ux.Logger.PrintToUser("Unsupported cmd: %s is not supported by local clusters", logging.LightBlue.Wrap(what))
 	return nil
-}
-
-func getRPCEndpoint(endpoint string, blockchainID string) string {
-	return models.NewDevnetNetwork(endpoint, 0).BlockchainEndpoint(blockchainID)
-}
-
-func getWSEndpoint(endpoint string, blockchainID string) string {
-	return models.NewDevnetNetwork(endpoint, 0).BlockchainWSEndpoint(blockchainID)
 }
