+++ pkg/networkoptions/network_options.go
@@ -27,6 +27,7 @@ const (
 	Mainnet
 	Fuji
 	Local
+	EtnaDevnet
 	Devnet
 	Cluster
 )
@@ -39,6 +40,8 @@ func (n NetworkOption) String() string {
 		return "Fuji Testnet"
 	case Local:
 		return "Local Network"
+	case EtnaDevnet:
+		return "Etna Devnet"
 	case Devnet:
 		return "Devnet"
 	case Cluster:
@@ -55,6 +58,8 @@ func NetworkOptionFromString(s string) NetworkOption {
 		return Fuji
 	case "Local Network":
 		return Local
+	case "Etna Devnet":
+		return EtnaDevnet
 	case "Devnet":
 		return Devnet
 	case "Cluster":
@@ -64,12 +69,13 @@ func NetworkOptionFromString(s string) NetworkOption {
 }
 
 type NetworkFlags struct {
-	UseLocal    bool
-	UseDevnet   bool
-	UseFuji     bool
-	UseMainnet  bool
-	Endpoint    string
-	ClusterName string
+	UseLocal      bool
+	UseEtnaDevnet bool
+	UseDevnet     bool
+	UseFuji       bool
+	UseMainnet    bool
+	Endpoint      string
+	ClusterName   string
 }
 
 func AddNetworkFlagsToCmd(cmd *cobra.Command, networkFlags *NetworkFlags, addEndpoint bool, supportedNetworkOptions []NetworkOption) {
@@ -78,6 +84,8 @@ func AddNetworkFlagsToCmd(cmd *cobra.Command, networkFlags *NetworkFlags, addEnd
 		switch networkOption {
 		case Local:
 			cmd.Flags().BoolVarP(&networkFlags.UseLocal, "local", "l", false, "operate on a local network")
+		case EtnaDevnet:
+			cmd.Flags().BoolVarP(&networkFlags.UseEtnaDevnet, "etna-devnet", "", false, "operate on an etna devnet network")
 		case Devnet:
 			cmd.Flags().BoolVar(&networkFlags.UseDevnet, "devnet", false, "operate on a devnet network")
 			addEndpoint = true
@@ -106,6 +114,8 @@ func GetNetworkFromSidecarNetworkName(
 	switch {
 	case networkName == models.Local.String():
 		return models.NewLocalNetwork(), nil
+	case networkName == models.EtnaDevnet.String():
+		return models.NewEtnaDevnetNetwork(), nil
 	case strings.HasPrefix(networkName, Cluster.String()):
 		parts := strings.Split(networkName, " ")
 		if len(parts) != 2 {
@@ -218,11 +228,12 @@ func GetNetworkFromCmdLineFlags(
 	}
 	// supported flags
 	networkFlagsMap := map[NetworkOption]string{
-		Local:   "--local",
-		Devnet:  "--devnet",
-		Fuji:    "--fuji/--testnet",
-		Mainnet: "--mainnet",
-		Cluster: "--cluster",
+		Local:      "--local",
+		EtnaDevnet: "--etna-devnet",
+		Devnet:     "--devnet",
+		Fuji:       "--fuji/--testnet",
+		Mainnet:    "--mainnet",
+		Cluster:    "--cluster",
 	}
 	supportedNetworksFlags := strings.Join(utils.Map(supportedNetworkOptions, func(n NetworkOption) string { return networkFlagsMap[n] }), ", ")
 	// received option
@@ -230,6 +241,8 @@ func GetNetworkFromCmdLineFlags(
 	switch {
 	case networkFlags.UseLocal:
 		networkOption = Local
+	case networkFlags.UseEtnaDevnet:
+		networkOption = EtnaDevnet
 	case networkFlags.UseDevnet:
 		networkOption = Devnet
 	case networkFlags.UseFuji:
@@ -256,7 +269,7 @@ func GetNetworkFromCmdLineFlags(
 		return models.UndefinedNetwork, errMsg
 	}
 	// mutual exclusion
-	if !flags.EnsureMutuallyExclusive([]bool{networkFlags.UseLocal, networkFlags.UseDevnet, networkFlags.UseFuji, networkFlags.UseMainnet, networkFlags.ClusterName != ""}) {
+	if !flags.EnsureMutuallyExclusive([]bool{networkFlags.UseLocal, networkFlags.UseEtnaDevnet, networkFlags.UseDevnet, networkFlags.UseFuji, networkFlags.UseMainnet, networkFlags.ClusterName != ""}) {
 		return models.UndefinedNetwork, fmt.Errorf("network flags %s are mutually exclusive", supportedNetworksFlags)
 	}
 
@@ -344,6 +357,8 @@ func GetNetworkFromCmdLineFlags(
 	switch networkOption {
 	case Local:
 		network = models.NewLocalNetwork()
+	case EtnaDevnet:
+		network = models.NewEtnaDevnetNetwork()
 	case Devnet:
 		networkID := uint32(0)
 		if networkFlags.Endpoint != "" {
